refactor: 프로젝트 기간 선택 컴포넌트 분리 파일의 내용이 너무 긿어서 가능한 최대한 분리
